Move prepareDocument to the Indexer service.
